Separated createScene function into it's own file
